+++ app/Http/Controllers/Controller.php
@@ -10,4 +10,22 @@ use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 class Controller extends BaseController
 {
     use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
+
+    public function successResponse(string $message, ?string $dataType, $data, int $code = 200): \Illuminate\Http\JsonResponse
+    {
+        return response()->json([
+            'status'  => 'success',
+            'message' => $message,
+            $dataType => $data,
+        ], $code);
+    }
+
+    public function errorResponse(string $message, ?string $dataType, $data, int $code = 400): \Illuminate\Http\JsonResponse
+    {
+        return response()->json([
+            'status'  => 'error',
+            'message' => $message,
+            $dataType => $data,
+        ], $code);
+    }
 }