+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 - Added `Quill` rich text editor on discussion section ([issue-165](https://github.com/iluminar/goodwork/issues/165))
 - Added Discussion create form
 - Added Discussion details modal on Discussion Board section ([issue-40](https://github.com/iluminar/goodwork/issues/40))
+- Added delete option for message, project, team and office
+- Added email notification on member add to project/team/office
+
+### Changed
+- Updated to Laravel 5.7
+- Updated docker compose to use `php:7.2-fpm` & `php:7.2-cli`
 
 ## v0.6 (2018-10-04)